+++ samplesheets/views_api.py
@@ -838,13 +838,13 @@ class ProjectIrodsFileListAPIView(SODARAPIBaseProjectMixin, APIView):
         path = irods_backend.get_sample_path(project)
         try:
             with irods_backend.get_session() as irods:
-                irods_data = irods_backend.get_objects(irods, path)
+                obj_list = irods_backend.get_objects(irods, path)
         except Exception as ex:
             return Response(
                 {'detail': '{}: {}'.format(IRODS_QUERY_ERROR_MSG, ex)},
                 status=status.HTTP_404_NOT_FOUND,
             )
-        return Response(irods_data, status=status.HTTP_200_OK)
+        return Response({'irods_data': obj_list}, status=status.HTTP_200_OK)
 
 
 # TODO: Temporary HACK, should be replaced by proper API view